requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb
python-dotenv
tiktoken
